SAM-78 Refactor
