+++ src/lib/InputDisplay/Operations/OpSettingsTools.svelte
@@ -2,7 +2,7 @@
   import type { OperationId } from '$lib/state/operation';
   import { type RenderPolicy } from '$lib/state/operation';
   import { appState } from '$lib/state/state.svelte';
-  import { setRenderPolicyCommand } from '$lib/state/undo';
+  import { setRenderPolicyCommand } from '$lib/state/undo/undo';
 
   export let operationId: OperationId;
   export let operationName: string;
